Move functions to error
